backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.3.2
mistralai==0.0.12
together==0.2.6
numpy==1.26.4
scikit-learn==1.4.0
pydantic==2.6.1
python-multipart==0.0.9
httpx==0.25.2
tenacity==8.2.3
sentence-transformers==2.5.1
aiofiles==23.2.1
psutil==5.9.8
PyPDF2==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
pdf2image==1.17.0
tiktoken==0.5.2
transformers>=4.30.0
torch>=2.0.0
langchain-text-splitters>=0.0.1
semantic-text-splitter>=0.8.0
spacy>=3.7.0
tokenizers>=0.15.0 
sentencepiece>=0.1.99
